pyproject.toml
[tool.hatch.metadata]
allow-direct-references = true
[project]
name = "airbyte-connector-builder-mcp"
dynamic = ["version"]
description = "MCP server for Airbyte connector building operations with AI ownership focus"
authors = [{name = "Airbyte", email = "contact@airbyte.io"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "airbyte-cdk>=6.61.6,<7.0",
    "fastmcp>=2.12.1,<3.0",
    "pydantic>=2.7.0,<3.0",
    "requests>=2.25.0,<3.0",
    "python-dotenv>=1.0.0,<2.0",
    "jsonschema>=4.0.0,<5.0",
    "pyyaml>=6.0.0,<7.0",
    "privatebin>=0.3.0",
]

[dependency-groups]
dev = [
    "deptry>=0.23.0",
    "pdoc>=15.0.4",
    "pyrefly>=0.25.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.4",
    "ty>=0.0.1a15",
    "mypy>=1.13.0",
    "coverage>=7.0.0",
    "poethepoet>=0.29.0",
    "types-requests>=2.25.0",
    "types-jsonschema>=4.0.0",
    "types-pyyaml>=6.0.0",
]

[project.scripts]
# Main CLI:
connector-builder-mcp = "connector_builder_mcp.server:main"

# CLI alias to support `uvx airbyte-connector-builder-mcp --help` and similar:
airbyte-connector-builder-mcp = "connector_builder_mcp.server:main"

[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[tool.uv-dynamic-versioning]
enable = true
vcs = "git"
style = "pep440"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.hatch.build.targets.wheel]
packages = ["connector_builder_mcp"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "requires_creds: marks tests as requiring credentials (deselected by default)",
    "slow: marks tests as slow (deselected by default)",
]

[tool.deptry]
exclude = [
    "examples/",
    ".venv/",
    "tests/",
]
